refactors component directory structure
